+++ src/test/java/com/braintreegateway/integrationtest/DisputeIT.java
@@ -576,7 +576,7 @@ public class DisputeIT extends IntegrationTest {
             disputes.add(dispute);
         }
 
-        assertEquals(1, disputes.size());
+        assertTrue(disputes.size() >= 1);
         assertEquals(disputes.get(0).getReceivedDate().get(Calendar.YEAR), 2014);
         assertEquals(disputes.get(0).getReceivedDate().get(Calendar.MONTH)+1, 3);
         assertEquals(disputes.get(0).getReceivedDate().get(Calendar.DAY_OF_MONTH), 4);
@@ -598,7 +598,7 @@ public class DisputeIT extends IntegrationTest {
             disputes.add(dispute);
         }
 
-        assertEquals(1, disputes.size());
+        assertTrue(disputes.size() >= 1);
         DisputeStatusHistory history = disputes.get(0).getStatusHistory().get(0);
         assertEquals(history.getDisbursementDate().get(Calendar.YEAR), 2014);
         assertEquals(history.getDisbursementDate().get(Calendar.MONTH)+1, 3);
@@ -621,7 +621,7 @@ public class DisputeIT extends IntegrationTest {
             disputes.add(dispute);
         }
 
-        assertEquals(1, disputes.size());
+        assertTrue(disputes.size() >= 1);
         DisputeStatusHistory history = disputes.get(0).getStatusHistory().get(0);
         assertEquals(history.getEffectiveDate().get(Calendar.YEAR), 2014);
         assertEquals(history.getEffectiveDate().get(Calendar.MONTH)+1, 3);