+++ src/test/java/com/braintreegateway/integrationtest/TransactionIT.java
@@ -276,6 +276,7 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
         assertEquals(Calendar.getInstance().get(Calendar.YEAR), transaction.getCreatedAt().get(Calendar.YEAR));
         assertEquals(Calendar.getInstance().get(Calendar.YEAR), transaction.getUpdatedAt().get(Calendar.YEAR));
         assertNotNull(transaction.getRetrievalReferenceNumber());
+        assertNull(transaction.getAcquirerReferenceNumber());
 
         CreditCard creditCard = transaction.getCreditCard();
         assertEquals("411111", creditCard.getBin());
@@ -1936,6 +1937,7 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
         assertTrue(result.isSuccess());
         Transaction transaction = result.getTarget();
         assertTrue(transaction.getRecurring());
+        assertTrue(transaction.isRecurring());
     }
 
     @Test
@@ -1951,6 +1953,7 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
         assertTrue(result.isSuccess());
         Transaction transaction = result.getTarget();
         assertTrue(transaction.getRecurring());
+        assertTrue(transaction.isRecurring());
     }
 
     @Test
@@ -1966,6 +1969,7 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
         assertTrue(result.isSuccess());
         Transaction transaction = result.getTarget();
         assertTrue(transaction.getRecurring());
+        assertTrue(transaction.isRecurring());
     }
 
     @Test
@@ -1981,6 +1985,7 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
         assertTrue(result.isSuccess());
         Transaction transaction = result.getTarget();
         assertFalse(transaction.getRecurring());
+        assertFalse(transaction.isRecurring());
     }
 
     @Test
@@ -1996,6 +2001,7 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
         assertTrue(result.isSuccess());
         Transaction transaction = result.getTarget();
         assertFalse(transaction.getRecurring());
+        assertFalse(transaction.isRecurring());
     }
 
     @Test
@@ -4415,6 +4421,13 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
         assertEquals(new BigDecimal("1000.00"), dispute.getAmount());
         assertEquals(new BigDecimal("1000.00"), dispute.getTransaction().getAmount());
         assertEquals("retrievaltransaction", dispute.getTransaction().getId());
+    }
+
+    @Test
+    public void findWithAcquirerReferenceNumber() throws Exception {
+        Transaction foundTransaction = gateway.transaction().find("transactionwithacquirerreferencenumber");
+
+        assertEquals("123456789 091019", foundTransaction.getAcquirerReferenceNumber());
     }
 
     @Test
@@ -7097,6 +7110,28 @@ public class TransactionIT extends IntegrationTest implements MerchantAccountTes
         assertNotNull(saleResult.getTarget().getPayPalDetails().getAuthorizationId());
         assertNotNull(saleResult.getTarget().getPayPalDetails().getToken());
         assertNotNull(saleResult.getTarget().getPayPalDetails().getDebugId());
+    }
+
+    @Test
+    public void createBillingAgreementPayPalTransactionAndAttemptToVault() {
+        String nonce = Nonce.PayPalBillingAgreement;
+        TransactionRequest request = new TransactionRequest().
+            amount(new BigDecimal("100.00")).
+            paymentMethodNonce(nonce).
+            options().
+                storeInVault(true).
+                done();
+
+        Result<Transaction> saleResult = gateway.transaction().sale(request);
+
+        assertTrue(saleResult.isSuccess());
+        assertNotNull(saleResult.getTarget().getPayPalDetails());
+        assertNotNull(saleResult.getTarget().getPayPalDetails().getPayerEmail());
+        assertNotNull(saleResult.getTarget().getPayPalDetails().getPaymentId());
+        assertNotNull(saleResult.getTarget().getPayPalDetails().getAuthorizationId());
+        assertNotNull(saleResult.getTarget().getPayPalDetails().getToken());
+        assertNotNull(saleResult.getTarget().getPayPalDetails().getDebugId());
+        assertNotNull(saleResult.getTarget().getPayPalDetails().getBillingAgreementId());
     }
 
     @Test