+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.3.0
+* Add `acquirerReferenceNumber` to `Transaction`
+* Add `billingAgreementId` to `PayPalDetails`
+* Add `isRecurring` to Transaction class
+* Deprecate `getRecurring` in Transaction class (use `isRecurring` instead)
+* Deprecate `recurring` in TransactionRequest
+
 ## 3.2.0
 * Deprecate `MasterpassCard` payment method
 * Deprecate `AmexExpressCheckoutCard` payment method